Template ProjectMortars on more specific things
Enabled by no longer needing the variables for the MortarDataHistory
type.  One function is now explicitly instantiated.
